Revert "reorganize vid events"
This reverts commit 42c1b0786eee416906625175a34b095dd41b9e70.
