+++ crates/task-impls/src/transactions.rs
@@ -13,8 +13,10 @@ use hotshot_task::{
     task_impls::HSTWithEvent,
 };
 use hotshot_types::{
+    block_impl::{NUM_CHUNKS, NUM_STORAGE_NODES},
     consensus::Consensus,
-    data::Leaf,
+    data::{test_srs, Leaf, VidDisperse, VidScheme, VidSchemeTrait},
+    message::Proposal,
     traits::{
         consensus_api::ConsensusApi,
         election::Membership,
@@ -27,10 +29,11 @@ use hotshot_utils::bincode::bincode_opts;
 use snafu::Snafu;
 use std::{
     collections::{HashMap, HashSet},
+    marker::PhantomData,
     sync::Arc,
     time::Instant,
 };
-use tracing::{debug, error, instrument, warn};
+use tracing::{debug, error, info, instrument, warn};
 
 /// A type alias for `HashMap<Commitment<T>, T>`
 type CommitmentMap<T> = HashMap<Commitment<T>, T>;
@@ -217,16 +220,56 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>, A: ConsensusApi<TYPES, I> + 
                             return None;
                         }
                     };
+                let encoded_txns = match payload.encode() {
+                    Ok(encoded) => encoded,
+                    Err(e) => {
+                        error!("Failed to encode the block payload: {:?}.", e);
+                        return None;
+                    }
+                };
+                // TODO <https://github.com/EspressoSystems/HotShot/issues/1686>
+                let srs = test_srs(NUM_STORAGE_NODES);
+                // TODO We are using constant numbers for now, but they will change as the quorum size
+                // changes.
+                // TODO <https://github.com/EspressoSystems/HotShot/issues/1693>
+                let vid = VidScheme::new(NUM_CHUNKS, NUM_STORAGE_NODES, &srs).unwrap();
+                let vid_disperse = vid
+                    .disperse(encoded_txns.into_iter().collect::<Vec<u8>>())
+                    .unwrap();
 
-                // Publish encoded transactions to the event stream
+                // TODO never clone a block
+                // https://github.com/EspressoSystems/HotShot/issues/1858
                 self.event_stream
-                    .publish(HotShotEvent::TransactionsSequenced(
-                        payload,
+                    .publish(HotShotEvent::BlockReady(
+                        payload.clone(),
                         metadata,
                         view + 1,
                     ))
                     .await;
 
+                // TODO (Keyao) Determine and update where to publish VidDisperseSend.
+                // <https://github.com/EspressoSystems/HotShot/issues/1817>
+                debug!("publishing VID disperse for view {}", *view + 1);
+                info!("New view: {}", *view);
+                self.event_stream
+                    .publish(HotShotEvent::VidDisperseSend(
+                        Proposal {
+                            data: VidDisperse {
+                                view_number: view + 1,
+                                payload_commitment: payload.commit(),
+                                shares: vid_disperse.shares,
+                                common: vid_disperse.common,
+                            },
+                            // TODO (Keyao) This is also signed in DA task.
+                            signature: TYPES::SignatureKey::sign(
+                                &self.private_key,
+                                payload.commit().as_ref(),
+                            ),
+                            _pd: PhantomData,
+                        },
+                        self.public_key.clone(),
+                    ))
+                    .await;
                 return None;
             }
             HotShotEvent::Shutdown => {
