+++ crates/testing/tests/network_task.rs
@@ -78,14 +78,9 @@ async fn test_network_task() {
     let mut output = HashMap::new();
 
     input.push(HotShotEvent::ViewChange(ViewNumber::new(1)));
-    input.push(HotShotEvent::TransactionsSequenced(
+    input.push(HotShotEvent::BlockReady(
         block.clone(),
         (),
-        ViewNumber::new(2),
-    ));
-    input.push(HotShotEvent::BlockReady(
-        block.commit(),
-        da_vid_disperse.clone().data,
         ViewNumber::new(2),
     ));
     input.push(HotShotEvent::DAProposalSend(da_proposal.clone(), pub_key));
@@ -106,19 +101,9 @@ async fn test_network_task() {
         2, // 2 occurrences: 1 from `input`, 1 from the DA task
     );
     output.insert(
-        HotShotEvent::TransactionsSequenced(block.clone(), (), ViewNumber::new(2)),
-        1,
-    );
-
-    output.insert(
-        HotShotEvent::BlockReady(
-            block.commit(),
-            da_vid_disperse.clone().data,
-            ViewNumber::new(2),
-        ),
+        HotShotEvent::BlockReady(block.clone(), (), ViewNumber::new(2)),
         2,
     );
-
     output.insert(
         HotShotEvent::VidDisperseRecv(da_vid_disperse.clone(), pub_key),
         1,
